shardtree: Move `ShardStore`-related types into module
