+++ shardtree/src/testing.rs
@@ -9,7 +9,7 @@ use proptest::sample::select;
 use incrementalmerkletree::{testing, Hashable};
 
 use super::*;
-use crate::memory::MemoryShardStore;
+use crate::store::{memory::MemoryShardStore, ShardStore};
 
 pub fn arb_retention_flags() -> impl Strategy<Value = RetentionFlags> + Clone {
     select(vec![
